services/python-service/pyproject.toml
[project]
name = "python-service"
version = "1.0.0"
description = "Python service with OpenTelemetry instrumentation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
